Measure time in microseconds
With a granularity of 4 microseconds.

+++ src/GcControllerLib/GcControllerLib.h
@@ -7,13 +7,13 @@ extern "C" {
   uint16_t blinkTime();
 }
 
-// You may only instantiate a single instance of this class globally.
+// Don't instantiate this class; instead, use the gc_controller instance already provided.
 class GcController {
 public:
     static void init();
 
-    // time since last pin-low interrupt, in 10ths of a millisecond
-    uint16_t time_since_pin_low();
+    // time since last pin-low interrupt, in microseconds, with a granularity of 4 microseconds
+    uint32_t time_since_pin_low();
 };
 
 extern GcController gc_controller;
